Add files via upload
Files from Sequoia repository - compiled by Matt Gregory.

Hey Keith,
I hacked a bit on the common library and got that to build, although I made some changes that might cause a few issues.  The main errors popping up had to do with a comparison operator for enumerations not being defined and I cast them to booleans and that made it happy (but not necessarily correct).  

But then once you start trying to compile SUE, you’re in for a world of hurt.  The message you got before about “The system cannot find the path specified” looks like it’s related to a pre-build step that requires perl to be installed for some of the parser library stuff.  I really think that Jimm might be the only one who would be able to step you through it all.

I can send you my changes which allowed the common lib to be built, but I’m not sure that gets you terribly far.

matt