Compiled version of SUE
Executable and test parameters from Jimm Domingo (summer 2016). This executable is from the Release folder (SUE\1.0\Release) with a file date of 17 July 2004 which is the most recent compiled version of SUE that I have found. None of the source files I found were newer than this, so I'm assuming it is the current version.